+++ CHANGELOG.md
@@ -1,8 +1,14 @@
+# 0.0.29
+
+## Feat
+
+- `extraPadding` added to ImpaktfullListView
+
 # 0.0.28
 
 ## Feat
 
-- Added margin & padding to the ImpaktfullCard
+- Added `margin` & `padding` to the ImpaktfullCard
 
 # 0.0.27